refactor: #90 음향 장비 package 분리
