+++ src/main/java/com/ajou/hertz/domain/instrument/repository/InstrumentRepositoryCustomImpl.java
@@ -2,7 +2,7 @@ package com.ajou.hertz.domain.instrument.repository;
 
 import static com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.entity.QAcousticAndClassicGuitar.*;
 import static com.ajou.hertz.domain.instrument.amplifier.entity.QAmplifier.*;
-import static com.ajou.hertz.domain.instrument.entity.QAudioEquipment.*;
+import static com.ajou.hertz.domain.instrument.audio_equipment.entity.QAudioEquipment.*;
 import static com.ajou.hertz.domain.instrument.bass_guitar.entity.QBassGuitar.*;
 import static com.ajou.hertz.domain.instrument.effector.entity.QEffector.*;
 import static com.ajou.hertz.domain.instrument.electric_guitar.entity.QElectricGuitar.*;
@@ -22,14 +22,14 @@ import com.ajou.hertz.domain.instrument.constant.InstrumentProgressStatus;
 import com.ajou.hertz.domain.instrument.constant.InstrumentSortOption;
 import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.dto.request.AcousticAndClassicGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.amplifier.dto.request.AmplifierFilterConditions;
-import com.ajou.hertz.domain.instrument.dto.request.AudioEquipmentFilterConditions;
+import com.ajou.hertz.domain.instrument.audio_equipment.dto.request.AudioEquipmentFilterConditions;
 import com.ajou.hertz.domain.instrument.bass_guitar.dto.request.BassGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.effector.dto.request.EffectorFilterConditions;
 import com.ajou.hertz.domain.instrument.electric_guitar.dto.request.ElectricGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.InstrumentFilterConditions;
 import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.entity.AcousticAndClassicGuitar;
 import com.ajou.hertz.domain.instrument.amplifier.entity.Amplifier;
-import com.ajou.hertz.domain.instrument.entity.AudioEquipment;
+import com.ajou.hertz.domain.instrument.audio_equipment.entity.AudioEquipment;
 import com.ajou.hertz.domain.instrument.bass_guitar.entity.BassGuitar;
 import com.ajou.hertz.domain.instrument.effector.entity.Effector;
 import com.ajou.hertz.domain.instrument.electric_guitar.entity.ElectricGuitar;
